+++ src/main/java/kono/ceu/materialreplication/common/machines/MRMetaTileEntities.java
@@ -20,12 +20,14 @@ import kono.ceu.materialreplication.common.machines.single.MetaTileEntityScrapMa
 
 public class MRMetaTileEntities {
 
-    // Single : 20000 - 200044
+    // Single Machine
+    // Default: 20000 - 200044
     public static SimpleMachineMetaTileEntity[] DECONSTRUCTOR = new SimpleMachineMetaTileEntity[GTValues.V.length - 1];
     public static SimpleMachineMetaTileEntity[] REPLICATOR = new SimpleMachineMetaTileEntity[GTValues.V.length - 1];
     public static MetaTileEntityScrapMaker[] SCRAPPER = new MetaTileEntityScrapMaker[GTValues.V.length - 1];
 
-    // Multi : 21000
+    // Multi Machine
+    // Default: 21000-
     public static MetaTileEntityLargeDeconstructor LARGE_DECONSTRUCTOR;
     public static MetaTileEntityLargeScrapper LARGE_SCRAPPER;
 
@@ -35,38 +37,44 @@ public class MRMetaTileEntities {
     }
 
     public static void registerSingleMachine() {
-        // 20000 - 20014 : Material Deconstructor
+        // Material Deconstructor
+        // Default: 20000 - 20014
+        int id = baseID;
         for (int i = 1; i < DECONSTRUCTOR.length; i++) {
             String voltageName = GTValues.VN[i].toLowerCase();
             if (!GregTechAPI.isHighTier() && i > GTValues.UV)
                 break;
             if (tierDeconstruct <= i)
-                DECONSTRUCTOR[i] = registerMetaTileEntity(20000 + (i - 1),
+                DECONSTRUCTOR[i] = registerMetaTileEntity(id + (i - 1),
                         new SimpleMachineMetaTileEntity(
                                 new ResourceLocation(MODID, (String.format("%s.%s", "deconstructor", voltageName))),
                                 MRRecipeMaps.DECONSTRUCTION_RECIPES, MRTextures.DECONSTRUCTOR_OVERLAY, i, true,
                                 GTUtility.hvCappedTankSizeFunction));
         }
 
-        // 20015 - 20029 : Material Replicator
+        // Material Replicator
+        // Default: 20015 - 20029
+        id = id + 15;
         for (int i = 1; i < REPLICATOR.length; i++) {
             String voltageName = GTValues.VN[i].toLowerCase();
             if (!GregTechAPI.isHighTier() && i > GTValues.UV)
                 break;
             if (tierReplicate <= i)
-                REPLICATOR[i] = registerMetaTileEntity(20015 + (i - 1),
+                REPLICATOR[i] = registerMetaTileEntity(id + (i - 1),
                         new SimpleMachineMetaTileEntity(
                                 new ResourceLocation(MODID, (String.format("%s.%s", "replicator", voltageName))),
                                 MRRecipeMaps.REPLICATION_RECIPES, MRTextures.REPLICATOR_OVERLAY, i, true,
                                 GTUtility.hvCappedTankSizeFunction));
         }
 
-        // 20030 - 20044 Scrapper
+        // Scrapper
+        // Default: 20030 - 20044
+        id = id + 15;
         for (int i = 0; i < SCRAPPER.length - 1; i++) {
             String voltageName = GTValues.VN[i + 1].toLowerCase();
             if (!GregTechAPI.isHighTier() && i > GTValues.UV)
                 break;
-            SCRAPPER[i] = registerMetaTileEntity(20030 + i,
+            SCRAPPER[i] = registerMetaTileEntity(id + i,
                     new MetaTileEntityScrapMaker(
                             new ResourceLocation(MODID, (String.format("%s.%s", "scrapmaker", voltageName))),
                             MRRecipeMaps.SCRAPMAKER_RECIPES, MRTextures.SCRAPPER_OVERLAY, i + 1,
@@ -76,12 +84,15 @@ public class MRMetaTileEntities {
     }
 
     public static void registerMultiMachine() {
-        // LargeDeconstructor 21000
-        LARGE_DECONSTRUCTOR = registerMetaTileEntity(21000,
+        int id = baseID + 1000;
+        // LargeDeconstructor
+        // Default: 21000
+        LARGE_DECONSTRUCTOR = registerMetaTileEntity(id,
                 new MetaTileEntityLargeDeconstructor(mrId("large_deconstructor")));
 
-        // LargeScrapper 21001
-        LARGE_SCRAPPER = registerMetaTileEntity(21001,
+        // LargeScrapper
+        // Default: 21001
+        LARGE_SCRAPPER = registerMetaTileEntity(id + 1,
                 new MetaTileEntityLargeScrapper(mrId("large_scrapper")));
     }